+++ src/program/ui/MainWindow.cpp
@@ -1291,14 +1291,14 @@ void MainWindow::applyLaunchTimeSettings()
 
     context->config.sc.initial_framerate_num = fpsNumField->value();
     context->config.sc.initial_framerate_den = fpsDenField->value();
-    if (! context->config.initial_monotonic_time_set_via_cli) {
+    if (! context->config.cli_overridden_keys.count("initial_monotonic_time_sec"))
         context->config.sc.initial_monotonic_time_sec = elapsedTimeSec->value();
+    if (! context->config.cli_overridden_keys.count("initial_monotonic_time_nsec"))
         context->config.sc.initial_monotonic_time_nsec = elapsedTimeNsec->value();
-    }
-    if (! context->config.initial_time_set_via_cli) {
+    if (! context->config.cli_overridden_keys.count("initial_time_sec"))
         context->config.sc.initial_time_sec = realTimeSec->value();
+    if (! context->config.cli_overridden_keys.count("initial_time_nsec"))
         context->config.sc.initial_time_nsec = realTimeNsec->value();
-    }
 }
 
 bool MainWindow::confirmDownloadsAllowed()