+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * Can tweak the number of cpu cores seen by the game
 * Vulkan: Use a fence to wait for each image to be presented (#741)
+* Add a command-line option to set any setting at startup
 
 ### Changed